fix compiler warnings.

+++ quickdialog/QDateTimeInlineElement.m
@@ -13,6 +13,7 @@
 //
 
 #import "QDateEntryTableViewCell.h"
+#import "QDateInlineTableViewCell.h"
 #import "ISO8601DateFormatter.h"
 #import "QDateTimeInlineElement.h"
 #import "QuickDialog.h"
@@ -219,7 +220,7 @@
     NSInteger seconds = ti % 60;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hours = (ti / 3600);
-    return [NSString stringWithFormat:@"%02i:%02i:%02i", hours, minutes, seconds];
+    return [NSString stringWithFormat:@"%02ld:%02ld:%02ld", (long)hours, (long)minutes, (long)seconds];
 }
